% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-aov.R
\name{aov-estimator}
\alias{aov-estimator}
\title{ANOVA-based Estimation Estimator for G-Studies}
\description{
Functions to fit G-study models using the method of moments (ANOVA-based
variance component estimation). This is an internal estimator used by \code{\link[=gstudy]{gstudy()}}.
}
\keyword{internal}
