% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends}
\alias{backends}
\title{Backend Management for G-Studies}
\description{
Functions to manage and interact with different fitting backends (lme4, brms).
These are internal functions used by \code{\link[=gstudy]{gstudy()}} and \code{\link[=dstudy]{dstudy()}}.
}
\keyword{internal}
