% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-helpers.R
\name{bf}
\alias{bf}
\title{Set up brms formulas}
\usage{
bf(formula, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{...}{Additional arguments passed to \code{\link[brms:bf]{brms::bf()}}.}
}
\value{
A brmsformula object.
}
\description{
This function is a wrapper around \code{\link[brms:bf]{brms::bf()}} to allow specifying
formulas for Bayesian models fit via the brms estimator in gstudy.
See \code{\link[brms:bf]{brms::bf()}} for full documentation.
}
