% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-helpers.R
\name{brms-formula-helpers}
\alias{brms-formula-helpers}
\title{Wrapper Functions for brms Formula Components}
\description{
These functions are wrappers around brms formula components to allow
specifying complex models (like multivariate ones) via the brms backend in gstudy.
}
