% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{build_facet_name_mapping}
\alias{build_facet_name_mapping}
\title{Build Mapping from ANOVA Source Names to User-Specified Names}
\usage{
build_facet_name_mapping(random_facet_specs, anova_results)
}
\description{
Build Mapping from ANOVA Source Names to User-Specified Names
}
\keyword{internal}
