% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{build_gstudy_draws_univariate}
\alias{build_gstudy_draws_univariate}
\title{Build Univariate gstudy Draws Tibble}
\usage{
build_gstudy_draws_univariate(object, vc_draws, n, n_draws, components)
}
\description{
Build Univariate gstudy Draws Tibble
}
\keyword{internal}
