% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-sizes.R
\name{calculate_residual_sample_size}
\alias{calculate_residual_sample_size}
\title{Calculate Sample Size for Residual}
\usage{
calculate_residual_sample_size(data, residual_facets)
}
\arguments{
\item{data}{A data frame.}

\item{residual_facets}{Character string of facets making up the residual,
separated by colons (e.g., "person:rater:item").}
}
\value{
Integer number of unique combinations.
}
\description{
Calculates the number of unique observations for the residual component.
}
\keyword{internal}
