% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-warnings.R
\name{check_estimation_issues}
\alias{check_estimation_issues}
\title{Check and Warn About Estimation Issues}
\usage{
check_estimation_issues(gstudy_obj)
}
\arguments{
\item{gstudy_obj}{A gstudy or mgstudy object.}
}
\description{
Called by dstudy() to check for estimation issues in the gstudy object
and issue appropriate warnings.
}
\keyword{internal}
