% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefficients}
\alias{coefficients}
\title{Generalizability and Dependability Coefficients}
\description{
Functions for computing G (generalizability) and D (dependability) coefficients
from variance components in generalizability theory.
}
\keyword{internal}
