% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-aov.R
\name{compute_aov_variance_components}
\alias{compute_aov_variance_components}
\title{Compute Variance Components from ANOVA Table using ANOVA-based Estimation}
\usage{
compute_aov_variance_components(anova_table, design_info)
}
\description{
Compute Variance Components from ANOVA Table using ANOVA-based Estimation
}
\keyword{internal}
