% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{compute_component_scale_factor}
\alias{compute_component_scale_factor}
\title{Compute scale factor for a variance component}
\usage{
compute_component_scale_factor(comp, n, object_spec, n_provided)
}
\description{
Compute scale factor for a variance component
}
\keyword{internal}
