% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite-coefficients.R
\name{compute_composite_error_variance}
\alias{compute_composite_error_variance}
\title{Compute Composite Error Variance}
\usage{
compute_composite_error_variance(
  vc,
  dimensions,
  weights,
  correlations,
  universe_spec,
  error_spec,
  aggregation,
  residual_is,
  error_type,
  n = NULL,
  object_spec = NULL
)
}
\arguments{
\item{vc}{A variance components tibble}

\item{dimensions}{Character vector of dimension names}

\item{weights}{Named numeric vector of weights}

\item{correlations}{List of covariance information from the G-study}

\item{universe_spec}{Character vector of universe components}

\item{error_spec}{Optional character vector of error components}

\item{aggregation}{Optional aggregation specification}

\item{residual_is}{Optional residual specification}

\item{error_type}{Either "relative" or "absolute"}

\item{scale_factor}{Factor by which to scale covariances}
}
\value{
The composite error variance as a numeric scalar
}
\description{
Calculates the composite relative or absolute error variance by summing
weighted variance-covariance matrices for all error components.
}
\keyword{internal}
