% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{compute_mom_multivariate_correlations}
\alias{compute_mom_multivariate_correlations}
\title{Compute Correlations for Multivariate Mom Model}
\usage{
compute_mom_multivariate_correlations(
  data,
  responses,
  random_facets,
  aov_models,
  aov_formulas,
  random_effect_cors = list(),
  compute_rescor = FALSE,
  compute_re_cors = FALSE
)
}
\arguments{
\item{data}{The data frame}

\item{responses}{Character vector of response variable names}

\item{random_facets}{Character vector of random effect facet names}

\item{aov_models}{List of fitted aov models (one per response variable)}

\item{random_effect_cors}{List of random effect correlation specifications}

\item{compute_rescor}{Whether to compute residual correlations}

\item{compute_re_cors}{Whether to compute random effect correlations}
}
\description{
Compute Correlations for Multivariate Mom Model
}
\keyword{internal}
