% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{compute_mom_variance_components}
\alias{compute_mom_variance_components}
\title{Compute Variance Components from ANOVA Table using Method of Moments}
\usage{
compute_mom_variance_components(anova_table, design_info)
}
\description{
Compute Variance Components from ANOVA Table using Method of Moments
}
\keyword{internal}
