% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{compute_sem}
\alias{compute_sem}
\title{Compute Standard Error of Measurement}
\usage{
compute_sem(error_variance)
}
\arguments{
\item{error_variance}{The error variance (relative or absolute).}
}
\value{
The standard error of measurement.
}
\description{
Calculates the standard error of measurement (SEM) for a D-study design.
}
\keyword{internal}
