% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable-optimization.R
\name{compute_var_for_weights}
\alias{compute_var_for_weights}
\title{Compute VAR for Given Weights}
\usage{
compute_var_for_weights(weights, dstudy_obj, type = "rel")
}
\arguments{
\item{weights}{Numeric vector of weights (will be normalized)}

\item{dstudy_obj}{A dstudy object}

\item{type}{"rel" for var_rel or "abs" for var_abs}
}
\value{
Named vector of VAR values for each subscale
}
\description{
Computes the Value Added Ratio for all subscales given a weight vector.
Used as the objective function for subscale optimization.
}
\keyword{internal}
