% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{default_prior}
\alias{default_prior}
\title{Get Default Priors for Bayesian Models}
\usage{
default_prior(object, ...)
}
\arguments{
\item{object}{An object whose class determines the method. Typically a formula,
brmsformula, or mvbrmsformula.}

\item{...}{Additional arguments passed to \code{\link[brms:default_prior]{brms::default_prior()}} including
\code{data} (a data frame containing the model data).}
}
\value{
A brmsprior object; a data frame containing information about all
parameters for which priors can be specified.
}
\description{
This function is a wrapper around \code{\link[brms:default_prior]{brms::default_prior()}} to allow
inspecting default priors that would be applied to a model.
See \code{\link[brms:default_prior]{brms::default_prior()}} for full documentation.
}
\examples{
\dontrun{
# Get default priors for a gstudy model
prior_info <- default_prior(score ~ (1 | person) + (1 | item), data = mydata)
print(prior_info)
}
}
\references{
Paul Bürkner (2017). brms: An R Package for Bayesian Multilevel Models.
\emph{Journal of Statistical Software}, 80(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
