% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{empty_prior}
\alias{empty_prior}
\title{Create an Empty Prior Object}
\usage{
empty_prior()
}
\value{
A brmsprior object with no rows.
}
\description{
This function is a wrapper around \code{\link[brms:set_prior]{brms::empty_prior()}} to create
an empty prior object to which priors can be added.
See \code{\link[brms:set_prior]{brms::empty_prior()}} for full documentation.
}
\examples{
\dontrun{
# Start with empty prior and add custom priors
my_prior <- empty_prior()
my_prior <- rbind(my_prior, set_prior("normal(0, 1)", class = "sd", group = "person"))
g <- gstudy(score ~ (1 | person), data = mydata, prior = my_prior, backend = "brms")
}
}
\references{
Paul Bürkner (2017). brms: An R Package for Bayesian Multilevel Models.
\emph{Journal of Statistical Software}, 80(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
