% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-warnings.R
\name{estimation-warnings}
\alias{estimation-warnings}
\title{Estimation Issue Detection Functions}
\description{
Functions to detect estimation issues from different estimators (lme4, mom, brms)
and issue appropriate warnings in dstudy.
}
\keyword{internal}
