% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimators}
\alias{estimators}
\title{Estimator Management for G-Studies}
\description{
Functions to manage and interact with different fitting estimators (lme4, brms).
These are internal functions used by \code{\link[=gstudy]{gstudy()}} and \code{\link[=dstudy]{dstudy()}}.
}
\keyword{internal}
