% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_draws.gstudy}
\alias{extract_draws.gstudy}
\title{Extract Draws from gstudy Objects}
\usage{
extract_draws.gstudy(object, ...)
}
\arguments{
\item{object}{A gstudy object.}

\item{...}{Additional arguments passed to brms::as_draws_matrix.}
}
\value{
A posterior draws_matrix object containing posterior draws for all model parameters.
}
\description{
Extract posterior draws from a gstudy object that was fit with the brms backend.
This is a wrapper around brms::as_draws_matrix.
}
