% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{extract_dstudy_coefficients}
\alias{extract_dstudy_coefficients}
\title{Extract Coefficient Draws from dstudy (Univariate)}
\usage{
extract_dstudy_coefficients(object, coefficients, dims)
}
\description{
Extract Coefficient Draws from dstudy (Univariate)
}
\keyword{internal}
