% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{extract_dstudy_composite}
\alias{extract_dstudy_composite}
\title{Extract Composite Posterior Draws from dstudy}
\usage{
extract_dstudy_composite(object, coefficients)
}
\description{
Extract Composite Posterior Draws from dstudy
}
\keyword{internal}
