% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{extract_dstudy_var_as_list}
\alias{extract_dstudy_var_as_list}
\title{Extract VAR Draws as Named List}
\usage{
extract_dstudy_var_as_list(object, dims)
}
\description{
Returns a named list of tibbles, one per dimension.
}
\keyword{internal}
