% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{extract_facets_from_formula}
\alias{extract_facets_from_formula}
\title{Extract Facet Names and Levels from Formula}
\usage{
extract_facets_from_formula(formula)
}
\arguments{
\item{formula}{An lme4-style formula}
}
\value{
Named list of facet levels (all set to 10 as default)
}
\description{
Extract Facet Names and Levels from Formula
}
\keyword{internal}
