% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{extract_mv_coefficients_as_list}
\alias{extract_mv_coefficients_as_list}
\title{Extract Multivariate Coefficient Draws as Named List}
\usage{
extract_mv_coefficients_as_list(posterior, coefficients, dims)
}
\description{
Returns a named list of tibbles (no dim or type columns).
}
\keyword{internal}
