% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{extract_sweep_coefficients}
\alias{extract_sweep_coefficients}
\title{Extract Sweep Coefficient Draws}
\usage{
extract_sweep_coefficients(object, coefficients, dims)
}
\description{
Extract Sweep Coefficient Draws
}
\keyword{internal}
