% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable.R
\name{extract_variance_draws_from_gstudy}
\alias{extract_variance_draws_from_gstudy}
\title{Extract Variance Draws from G-Study for Viable Recalculation}
\usage{
extract_variance_draws_from_gstudy(gstudy_obj, n_draws = 1000)
}
\arguments{
\item{gstudy_obj}{A gstudy object}

\item{n_draws}{Number of pseudo-draws to generate for mom backend}
}
\description{
Extract Variance Draws from G-Study for Viable Recalculation
}
\keyword{internal}
