% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-package.R
\docType{package}
\name{facet-package}
\alias{facet}
\alias{facet-package}
\title{facet: facet: Univeriate and Multivariate Generalizability Theory with frequentist and Bayesian methods}
\description{
Conduct generalizability theory analyses using variance component models estimated with various methods (including ANOVA method of momnets, REML with lme4, and NUTES with brms. Provides functions for G-studies (estimating variance components) and D-studies (computing generalizability and dependability coefficients).
}
\details{
The facet package provides tools for conducting generalizability theory analyses
using variance component models. It supports multiple backends including lme4
for frequentist analyses and brms for Bayesian analyses.

Key functions:
\itemize{
\item \code{\link[=gstudy]{gstudy()}}: Conduct a G-study to estimate variance components
\item \code{\link[=dstudy]{dstudy()}}: Conduct a D-study to compute reliability coefficients
}
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/yourusername/mgt}
}
}
\author{
\strong{Maintainer}: George Michaelides \email{g.michaelides@uea.ac.uk}

Authors:
\itemize{
  \item George Michaelides \email{g.michaelides@uea.ac.uk}
  \item Duncan Jackson \email{duncan.jackson@kcl.ac.uk}
}

}
\keyword{internal}
