% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{find_residual_ms}
\alias{find_residual_ms}
\title{Find Residual MS for a Given Source}
\usage{
find_residual_ms(anova_results, source)
}
\description{
For nested designs, the residual for a given effect is the next interaction
that contains all the factors in the effect plus one more factor.
}
\keyword{internal}
