% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-aov.R
\name{fit_aov_multivariate}
\alias{fit_aov_multivariate}
\title{Fit Multivariate Model Using ANOVA-based Estimation}
\usage{
fit_aov_multivariate(formula, data, unbalanced = FALSE, nested = NULL, ...)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A data frame containing the variables}

\item{unbalanced}{Logical. If TRUE, uses Henderson's Method III for unbalanced designs}

\item{...}{Additional arguments}
}
\description{
Fit Multivariate Model Using ANOVA-based Estimation
}
\keyword{internal}
