% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{fit_brms}
\alias{fit_brms}
\title{Fit a Model Using brms}
\usage{
fit_brms(formula, data, prior = NULL, ...)
}
\arguments{
\item{formula}{A formula for the model (can be brmsformula).}

\item{data}{A data frame.}

\item{prior}{A brmsprior object or list of priors as created by \code{\link[=set_prior]{set_prior()}}
or related functions. If NULL, default priors are used.}

\item{...}{Additional arguments passed to \code{\link[brms:brm]{brms::brm()}}.}
}
\value{
A fitted brmsfit object.
}
\description{
Fits a Bayesian mixed effects model using brms::brm(). Handles both
univariate and multivariate formulas.
}
\keyword{internal}
