% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{fit_mom_multivariate}
\alias{fit_mom_multivariate}
\title{Fit Multivariate Model Using Method of Moments}
\usage{
fit_mom_multivariate(formula, data, unbalanced = FALSE, ...)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A data frame containing the variables}

\item{unbalanced}{Logical. If TRUE, uses Henderson's Method III for unbalanced designs}

\item{...}{Additional arguments}
}
\description{
Fit Multivariate Model Using Method of Moments
}
\keyword{internal}
