% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_correlated_effects}
\alias{generate_correlated_effects}
\title{Generate Correlated Effects Using Cholesky Decomposition}
\usage{
generate_correlated_effects(n, cor_matrix, sds)
}
\arguments{
\item{n}{Number of observations}

\item{cor_matrix}{Correlation matrix}

\item{sds}{Standard deviations (vector of length ncol(cor_matrix))}
}
\value{
Matrix of correlated effects (n x ncol(cor_matrix))
}
\description{
Generate Correlated Effects Using Cholesky Decomposition
}
\keyword{internal}
