% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_design_structure}
\alias{generate_design_structure}
\title{Generate Design Structure}
\usage{
generate_design_structure(facets, nested = NULL, prefix = "")
}
\arguments{
\item{facets}{Named list of facet levels}

\item{nested}{Named list of nesting relationships}

\item{prefix}{Prefix for level names}
}
\value{
Data frame with facet columns
}
\description{
Generate Design Structure
}
\keyword{internal}
