% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_multivariate_long}
\alias{generate_multivariate_long}
\title{Generate Long-Format Multivariate Data}
\usage{
generate_multivariate_long(
  facets,
  vc,
  n_dims,
  dim_names,
  dim_var,
  response_var,
  sd_residual,
  residual_cor,
  re_cor,
  nested,
  prefix
)
}
\arguments{
\item{facets}{Named list of facet levels}

\item{vc}{Named list of variance components}

\item{n_dims}{Number of dimensions}

\item{dim_names}{Names of dimensions}

\item{dim_var}{Name of dimension indicator variable}

\item{response_var}{Name of response variable}

\item{sd_residual}{Residual standard deviation (or vector for heterogeneous)}

\item{residual_cor}{Residual correlation matrix}

\item{re_cor}{Named list of random effect correlation matrices}

\item{nested}{Named list of nesting relationships}

\item{prefix}{Prefix for level names}
}
\value{
Data frame with simulated multivariate data in long format
}
\description{
Generate Long-Format Multivariate Data
}
\keyword{internal}
