% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_univariate}
\alias{generate_univariate}
\title{Generate Univariate Data}
\usage{
generate_univariate(
  facets,
  vc,
  sd_residual,
  response_var,
  nested = NULL,
  prefix = ""
)
}
\arguments{
\item{facets}{Named list of facet levels}

\item{vc}{Named list of variance components}

\item{sd_residual}{Residual standard deviation}

\item{response_var}{Name of response variable}

\item{nested}{Named list of nesting relationships}

\item{prefix}{Prefix for level names}
}
\value{
Data frame with simulated data
}
\description{
Generate Univariate Data
}
\keyword{internal}
