% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{glance.gstudy}
\alias{glance.gstudy}
\alias{glance.mgstudy}
\title{Glance Method for gstudy Objects}
\usage{
\method{glance}{gstudy}(x, ...)

\method{glance}{mgstudy}(x, ...)
}
\arguments{
\item{x}{An mgstudy object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A one-row tibble.

A one-row tibble.
}
\description{
Returns a one-row summary of model statistics.

Returns a one-row summary of model statistics for multivariate G-studies.
}
