% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-draws.R
\name{gt_draws_dstudy_multivariate}
\alias{gt_draws_dstudy_multivariate}
\title{Handle Multivariate dstudy Draws}
\usage{
gt_draws_dstudy_multivariate(object, what, coefficients, dims)
}
\description{
Handle Multivariate dstudy Draws
}
\keyword{internal}
