% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.dstudy}
\alias{is.dstudy}
\title{Check if Object is a dstudy Object}
\usage{
is.dstudy(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if x is a dstudy object, FALSE otherwise.
}
\description{
Check if Object is a dstudy Object
}
