% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.gstudy}
\alias{is.gstudy}
\title{Check if Object is a gstudy Object}
\usage{
is.gstudy(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if x is a gstudy object, FALSE otherwise.
}
\description{
Check if Object is a gstudy Object
}
