% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.mgstudy}
\alias{is.mgstudy}
\title{Check if Object is an mgstudy Object}
\usage{
is.mgstudy(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if x is an mgstudy object, FALSE otherwise.
}
\description{
Check if Object is an mgstudy Object
}
