% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{is_object_component}
\alias{is_object_component}
\title{Check if Component is Part of Object Specification}
\usage{
is_object_component(component, object_spec)
}
\arguments{
\item{component}{Character string naming the variance component.}

\item{object_spec}{Character vector of object component names.}
}
\value{
TRUE if the component is part of the object specification.
}
\description{
Determines whether a variance component is part of the object of measurement
specification (which can include multiple components).
}
\keyword{internal}
