% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mgt-classes}
\alias{mgt-classes}
\title{S3 Class Definitions for mgt Package}
\description{
This file defines the S3 classes used throughout the mgt package.
The main classes are "gstudy" for G-study results and "dstudy" for D-study results.
}
\keyword{internal}
