% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-helpers.R
\name{mvbind}
\alias{mvbind}
\title{Bind response variables for multivariate models}
\usage{
mvbind(...)
}
\arguments{
\item{...}{Unquoted names of variables to bind.}
}
\value{
A matrix-like object for brms formulas.
}
\description{
This function is a wrapper around \code{\link[brms:mvbind]{brms::mvbind()}} to allow specifying
multivariate models via the brms backend in gstudy.
See \code{\link[brms:mvbind]{brms::mvbind()}} for full documentation.
}
