% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable-optimization.R
\name{optimize_composite_weights_internal}
\alias{optimize_composite_weights_internal}
\title{Optimize Weights for Composite Reliability}
\usage{
optimize_composite_weights_internal(dstudy_obj, optimize_target = "rel")
}
\arguments{
\item{dstudy_obj}{A dstudy object}

\item{optimize_target}{"rel" or "abs"}
}
\value{
List with optimization results
}
\description{
Main function for "composite" optimization method.
}
\keyword{internal}
