% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable-optimization.R
\name{optimize_subscale_weights_internal}
\alias{optimize_subscale_weights_internal}
\title{Optimize Weights for Subscale VAR}
\usage{
optimize_subscale_weights_internal(
  dstudy_obj,
  subscale = NULL,
  optimize_target = "rel"
)
}
\arguments{
\item{dstudy_obj}{A dstudy object}

\item{subscale}{Target subscale name (NULL for all)}

\item{optimize_target}{"rel" or "abs"}
}
\value{
List with optimization results
}
\description{
Main function for "subscale" optimization method.
}
\keyword{internal}
