% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable-optimization.R
\name{optimize_tuning_weights_internal}
\alias{optimize_tuning_weights_internal}
\title{Optimize Weights via Grid Search (Tuning)}
\usage{
optimize_tuning_weights_internal(
  dstudy_obj,
  grid_resolution = 0.1,
  optimize_target = "rel"
)
}
\arguments{
\item{dstudy_obj}{A dstudy object}

\item{grid_resolution}{Step size for grid}

\item{optimize_target}{"rel" or "abs"}
}
\value{
List with optimization results
}
\description{
Main function for "tuning" optimization method.
}
\keyword{internal}
