% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pairs.gstudy}
\alias{pairs.gstudy}
\title{Pairs Method for gstudy Objects}
\usage{
\method{pairs}{gstudy}(x, ...)
}
\arguments{
\item{x}{A gstudy object.}

\item{...}{Additional arguments passed to pairs.brmsfit.}
}
\value{
The result of pairs.brmsfit.
}
\description{
Wrapper for pairs.brmsfit to visualize parameter pairs from a gstudy
object that was fit with the brms backend.
}
