% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior-functions}
\alias{prior-functions}
\title{Prior Wrapper Functions for brms}
\description{
These functions are wrappers around brms prior functions to allow
specifying priors for Bayesian models fit via the brms backend in gstudy.
}
