% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior}
\alias{prior}
\title{Define Priors Using Tidy Evaluation}
\usage{
prior(prior, ...)
}
\arguments{
\item{prior}{A call or expression defining a distribution.}

\item{...}{Additional arguments passed to \code{\link[brms:set_prior]{brms::prior()}}.}
}
\value{
A brmsprior object.
}
\description{
This function is a wrapper around \code{\link[brms:set_prior]{brms::prior()}} to allow
specifying priors using tidy evaluation syntax.
See \code{\link[brms:set_prior]{brms::prior()}} for full documentation.
}
\examples{
\dontrun{
prior <- prior(normal(0, 1), class = sd)
g <- gstudy(score ~ (1 | person), data = mydata, prior = prior, backend = "brms")
}
}
\references{
Paul Bürkner (2017). brms: An R Package for Bayesian Multilevel Models.
\emph{Journal of Statistical Software}, 80(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
