% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_}
\alias{prior_}
\title{Define Priors Using Formula Syntax}
\usage{
prior_(prior, ...)
}
\arguments{
\item{prior}{A one-sided formula defining the prior distribution.}

\item{...}{Additional arguments passed to \code{\link[brms:set_prior]{brms::prior_()}}.}
}
\value{
A brmsprior object.
}
\description{
This function is a wrapper around \code{\link[brms:set_prior]{brms::prior_()}} to allow
specifying priors using formula syntax (with \code{~}).
See \code{\link[brms:set_prior]{brms::prior_()}} for full documentation.
}
\examples{
\dontrun{
prior <- prior_(~ normal(0, 1), class = "sd")
g <- gstudy(score ~ (1 | person), data = mydata, prior = prior, backend = "brms")
}
}
\references{
Paul Bürkner (2017). brms: An R Package for Bayesian Multilevel Models.
\emph{Journal of Statistical Software}, 80(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
