% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_string}
\alias{prior_string}
\title{Define Priors Using Character Strings}
\usage{
prior_string(prior, ...)
}
\arguments{
\item{prior}{A character string defining the prior distribution.}

\item{...}{Additional arguments passed to \code{\link[brms:set_prior]{brms::prior_string()}}.}
}
\value{
A brmsprior object.
}
\description{
This function is a wrapper around \code{\link[brms:set_prior]{brms::prior_string()}} to allow
specifying priors using character strings.
See \code{\link[brms:set_prior]{brms::prior_string()}} for full documentation.
}
\examples{
\dontrun{
prior <- prior_string("normal(0, 1)", class = "sd", group = "person")
g <- gstudy(score ~ (1 | person), data = mydata, prior = prior, backend = "brms")
}
}
\references{
Paul Bürkner (2017). brms: An R Package for Bayesian Multilevel Models.
\emph{Journal of Statistical Software}, 80(1), 1-28.
\doi{10.18637/jss.v080.i01}
}
