% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rajaratnam}
\alias{rajaratnam}
\title{Rajaratnam, Cronbach, & Gleser (1965) Example Data}
\format{
A data frame with 64 rows and 4 variables:
\describe{
\item{Person}{Factor with 8 levels indicating the person}
\item{Subtest}{Factor with 3 levels indicating the subtest}
\item{Item}{Factor with 4 levels indicating the item within subtest}
\item{Score}{Integer score}
}
}
\source{
Rajaratnam, N., Cronbach, L. J., & Gleser, G. C. (1965).
\emph{Generalizability of stratified-parallel tests.} Psychometrika, 30(1), 39-56.
}
\usage{
rajaratnam
}
\description{
A classic dataset illustrating a
Person crossed with Subtest and Item design.
}
\keyword{datasets}
