% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ranef}
\alias{ranef}
\title{Random Effects Extractor Generic}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{An object from which to extract random effects.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Random effects from the fitted model.
}
\description{
Generic function for extracting random effects (BLUPs/conditional modes) from fitted models.
Dispatches to the appropriate method based on the class of the object.
}
