% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-helpers.R
\name{set_rescor}
\alias{set_rescor}
\title{Set residual correlation between response variables}
\usage{
set_rescor(rescor)
}
\arguments{
\item{rescor}{Logical; whether to estimate residual correlations.}
}
\value{
A brmsformula helper object.
}
\description{
This function is a wrapper around \code{\link[brms:brmsformula-helpers]{brms::set_rescor()}} to allow specifying
residual correlations in multivariate models fit via the brms backend.
See \code{\link[brms:brmsformula-helpers]{brms::set_rescor()}} for full documentation.
}
