% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_dstudy}
\alias{validate_dstudy}
\title{Validate a dstudy Object}
\usage{
validate_dstudy(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
TRUE if valid, otherwise raises an error.
}
\description{
Checks that a dstudy object has the required components and they are valid.
}
\keyword{internal}
