% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_gstudy}
\alias{validate_gstudy}
\title{Validate a gstudy Object}
\usage{
validate_gstudy(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
TRUE if valid, otherwise raises an error.
}
\description{
Checks that a gstudy object has the required components and they are valid.
}
\keyword{internal}
