% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_mgstudy}
\alias{validate_mgstudy}
\title{Validate an mgstudy Object}
\usage{
validate_mgstudy(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
TRUE if valid, otherwise raises an error.
}
\description{
Checks that an mgstudy object has the required components and they are valid.
}
\keyword{internal}
