% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brennan}
\alias{brennan}
\title{Brennan's (2001) Generalizability Theory Example Data}
\format{
A data frame with 120 rows and 5 variables:
\describe{
\item{Task}{Factor with 3 levels indicating the task}
\item{Person}{Factor with 10 levels indicating the person/examinee}
\item{Rater}{Factor with 12 levels indicating the rater}
\item{Score}{Integer score}
\item{y}{Integer score (identical to Score)}
}
}
\source{
Brennan, R. L. (2001). \emph{Generalizability theory}. Springer-Verlag.
}
\usage{
brennan
}
\description{
A classic dataset from Brennan (2001) illustrating a Person crossed with Task
and Rater design (p x t x r). Commonly used to demonstrate complex
generalizability theory analyses.
}
\keyword{datasets}
