% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{compute_mom_vc_from_results}
\alias{compute_mom_vc_from_results}
\title{Compute Variance Components from ANOVA Results}
\usage{
compute_mom_vc_from_results(
  anova_results,
  data,
  response,
  random_facets,
  random_facet_specs = NULL
)
}
\description{
Compute Variance Components from ANOVA Results
}
\keyword{internal}
