% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mom.R
\name{fit_mom_multivariate}
\alias{fit_mom_multivariate}
\title{Fit Multivariate Model Using Method of Moments}
\usage{
fit_mom_multivariate(formula, data, ...)
}
\description{
Fit Multivariate Model Using Method of Moments
}
\keyword{internal}
