% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{glance.dstudy}
\alias{glance.dstudy}
\title{Glance Method for dstudy Objects}
\usage{
\method{glance}{dstudy}(x, ...)
}
\arguments{
\item{x}{A dstudy object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A one-row tibble.
}
\description{
Returns a one-row summary of D-study results.
}
