% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viable.R
\name{recalculate_var_with_weights}
\alias{recalculate_var_with_weights}
\title{Recalculate VAR with Alternative Weights}
\usage{
recalculate_var_with_weights(dstudy_obj, weights, dims, ci, probs)
}
\description{
Recalculate VAR with Alternative Weights
}
\keyword{internal}
