% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.dstudy}
\alias{tidy.dstudy}
\title{Tidy Method for dstudy Objects}
\usage{
\method{tidy}{dstudy}(x, ...)
}
\arguments{
\item{x}{A dstudy object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A tibble.
}
\description{
Returns a tidy tibble of coefficients and variance components.
}
