% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.gstudy}
\alias{tidy.gstudy}
\alias{tidy.mgstudy}
\title{Tidy Method for gstudy Objects}
\usage{
\method{tidy}{gstudy}(x, ...)

\method{tidy}{mgstudy}(x, ...)
}
\arguments{
\item{x}{An mgstudy object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A tibble.

A tibble.
}
\description{
Returns a tidy tibble of variance components.

Returns a tidy tibble of variance components for multivariate G-studies.
}
